@@ -3,7 +3,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -11,16 +11,16 @@ dependencies:
   intl: ^0.16.0
   url_launcher: 5.4.2
   get_it: 4.0.1
-  auto_route: 0.6.9
+  auto_route:
   cupertino_icons: ^0.1.3
   connectivity: 0.4.8+2
   cached_network_image: ^2.2.0
-  json_annotation: 3.1.0
+  json_annotation: ^4.0.0-nullsafety.0
   shimmer: 1.1.1
   package_info: ^0.4.0+18
   fluttertoast: ^4.0.0
   smooth_page_indicator: 0.1.5
-  freezed: 0.12.2
+  freezed:  0.14.0-nullsafety.0
   flutter_bloc: 6.0.0
   cloud_firestore: 0.13.7
   firebase_crashlytics: 0.1.3+3
@@ -32,9 +32,9 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: 1.10.3
-  json_serializable: ^3.5.0
-  auto_route_generator: 0.6.10
+  build_runner:
+  json_serializable:
+  auto_route_generator:
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
@@ -43,7 +43,6 @@ flutter:
   uses-material-design: true
 
   assets:
-    #Images
     - assets/images/app_logo.png
 
 
