@@ -11,10 +11,10 @@ abstract class ResultState<T> with _$ResultState<T> {
 
   const factory ResultState.reLoading() = ReLoading<T>;
 
-  const factory ResultState.data({@required T data}) = Data<T>;
+  const factory ResultState.data({required T data}) = Data<T>;
 
-  const factory ResultState.error({@required String error}) = Error<T>;
+  const factory ResultState.error({required String error}) = Error<T>;
 
   const factory ResultState.unNotifiedError(
-      {@required T data, @required String error}) = UnNotifiedError<T>;
+      {required T data, required String error}) = UnNotifiedError<T>;
 }
