@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'my_address_state.dart';
 
@@ -9,23 +9,23 @@ part of 'my_address_state.dart';
 
 T _$identity<T>(T value) => value;
 
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 /// @nodoc
 class _$MyAddressStateTearOff {
   const _$MyAddressStateTearOff();
 
-// ignore: unused_element
   Loading loading() {
     return const Loading();
   }
 
-// ignore: unused_element
   ShowAccountDetails showAccountDetails(AccountDetails accountDetails) {
     return ShowAccountDetails(
       accountDetails,
     );
   }
 
-// ignore: unused_element
   Error error(String errorMessage) {
     return Error(
       errorMessage,
@@ -34,37 +34,40 @@ class _$MyAddressStateTearOff {
 }
 
 /// @nodoc
-// ignore: unused_element
 const $MyAddressState = _$MyAddressStateTearOff();
 
 /// @nodoc
 mixin _$MyAddressState {
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result loading(),
-    @required Result showAccountDetails(AccountDetails accountDetails),
-    @required Result error(String errorMessage),
-  });
+  TResult when<TResult extends Object?>({
+    required TResult Function() loading,
+    required TResult Function(AccountDetails accountDetails) showAccountDetails,
+    required TResult Function(String errorMessage) error,
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result loading(),
-    Result showAccountDetails(AccountDetails accountDetails),
-    Result error(String errorMessage),
-    @required Result orElse(),
-  });
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? loading,
+    TResult Function(AccountDetails accountDetails)? showAccountDetails,
+    TResult Function(String errorMessage)? error,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result loading(Loading value),
-    @required Result showAccountDetails(ShowAccountDetails value),
-    @required Result error(Error value),
-  });
+  TResult map<TResult extends Object?>({
+    required TResult Function(Loading value) loading,
+    required TResult Function(ShowAccountDetails value) showAccountDetails,
+    required TResult Function(Error value) error,
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result loading(Loading value),
-    Result showAccountDetails(ShowAccountDetails value),
-    Result error(Error value),
-    @required Result orElse(),
-  });
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Loading value)? loading,
+    TResult Function(ShowAccountDetails value)? showAccountDetails,
+    TResult Function(Error value)? error,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -119,26 +122,22 @@ class _$Loading implements Loading {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result loading(),
-    @required Result showAccountDetails(AccountDetails accountDetails),
-    @required Result error(String errorMessage),
+  TResult when<TResult extends Object?>({
+    required TResult Function() loading,
+    required TResult Function(AccountDetails accountDetails) showAccountDetails,
+    required TResult Function(String errorMessage) error,
   }) {
-    assert(loading != null);
-    assert(showAccountDetails != null);
-    assert(error != null);
     return loading();
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result loading(),
-    Result showAccountDetails(AccountDetails accountDetails),
-    Result error(String errorMessage),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? loading,
+    TResult Function(AccountDetails accountDetails)? showAccountDetails,
+    TResult Function(String errorMessage)? error,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (loading != null) {
       return loading();
     }
@@ -147,26 +146,22 @@ class _$Loading implements Loading {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result loading(Loading value),
-    @required Result showAccountDetails(ShowAccountDetails value),
-    @required Result error(Error value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Loading value) loading,
+    required TResult Function(ShowAccountDetails value) showAccountDetails,
+    required TResult Function(Error value) error,
   }) {
-    assert(loading != null);
-    assert(showAccountDetails != null);
-    assert(error != null);
     return loading(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result loading(Loading value),
-    Result showAccountDetails(ShowAccountDetails value),
-    Result error(Error value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Loading value)? loading,
+    TResult Function(ShowAccountDetails value)? showAccountDetails,
+    TResult Function(Error value)? error,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (loading != null) {
       return loading(this);
     }
@@ -199,7 +194,7 @@ class _$ShowAccountDetailsCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object accountDetails = freezed,
+    Object? accountDetails = freezed,
   }) {
     return _then(ShowAccountDetails(
       accountDetails == freezed
@@ -211,8 +206,7 @@ class _$ShowAccountDetailsCopyWithImpl<$Res>
 
 /// @nodoc
 class _$ShowAccountDetails implements ShowAccountDetails {
-  const _$ShowAccountDetails(this.accountDetails)
-      : assert(accountDetails != null);
+  const _$ShowAccountDetails(this.accountDetails);
 
   @override
   final AccountDetails accountDetails;
@@ -236,32 +230,29 @@ class _$ShowAccountDetails implements ShowAccountDetails {
       runtimeType.hashCode ^
       const DeepCollectionEquality().hash(accountDetails);
 
+  @JsonKey(ignore: true)
   @override
   $ShowAccountDetailsCopyWith<ShowAccountDetails> get copyWith =>
       _$ShowAccountDetailsCopyWithImpl<ShowAccountDetails>(this, _$identity);
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result loading(),
-    @required Result showAccountDetails(AccountDetails accountDetails),
-    @required Result error(String errorMessage),
+  TResult when<TResult extends Object?>({
+    required TResult Function() loading,
+    required TResult Function(AccountDetails accountDetails) showAccountDetails,
+    required TResult Function(String errorMessage) error,
   }) {
-    assert(loading != null);
-    assert(showAccountDetails != null);
-    assert(error != null);
     return showAccountDetails(accountDetails);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result loading(),
-    Result showAccountDetails(AccountDetails accountDetails),
-    Result error(String errorMessage),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? loading,
+    TResult Function(AccountDetails accountDetails)? showAccountDetails,
+    TResult Function(String errorMessage)? error,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (showAccountDetails != null) {
       return showAccountDetails(accountDetails);
     }
@@ -270,26 +261,22 @@ class _$ShowAccountDetails implements ShowAccountDetails {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result loading(Loading value),
-    @required Result showAccountDetails(ShowAccountDetails value),
-    @required Result error(Error value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Loading value) loading,
+    required TResult Function(ShowAccountDetails value) showAccountDetails,
+    required TResult Function(Error value) error,
   }) {
-    assert(loading != null);
-    assert(showAccountDetails != null);
-    assert(error != null);
     return showAccountDetails(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result loading(Loading value),
-    Result showAccountDetails(ShowAccountDetails value),
-    Result error(Error value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Loading value)? loading,
+    TResult Function(ShowAccountDetails value)? showAccountDetails,
+    TResult Function(Error value)? error,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (showAccountDetails != null) {
       return showAccountDetails(this);
     }
@@ -301,8 +288,10 @@ abstract class ShowAccountDetails implements MyAddressState {
   const factory ShowAccountDetails(AccountDetails accountDetails) =
       _$ShowAccountDetails;
 
-  AccountDetails get accountDetails;
-  $ShowAccountDetailsCopyWith<ShowAccountDetails> get copyWith;
+  AccountDetails get accountDetails => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $ShowAccountDetailsCopyWith<ShowAccountDetails> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -323,7 +312,7 @@ class _$ErrorCopyWithImpl<$Res> extends _$MyAddressStateCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object errorMessage = freezed,
+    Object? errorMessage = freezed,
   }) {
     return _then(Error(
       errorMessage == freezed ? _value.errorMessage : errorMessage as String,
@@ -333,7 +322,7 @@ class _$ErrorCopyWithImpl<$Res> extends _$MyAddressStateCopyWithImpl<$Res>
 
 /// @nodoc
 class _$Error implements Error {
-  const _$Error(this.errorMessage) : assert(errorMessage != null);
+  const _$Error(this.errorMessage);
 
   @override
   final String errorMessage;
@@ -356,32 +345,29 @@ class _$Error implements Error {
   int get hashCode =>
       runtimeType.hashCode ^ const DeepCollectionEquality().hash(errorMessage);
 
+  @JsonKey(ignore: true)
   @override
   $ErrorCopyWith<Error> get copyWith =>
       _$ErrorCopyWithImpl<Error>(this, _$identity);
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result loading(),
-    @required Result showAccountDetails(AccountDetails accountDetails),
-    @required Result error(String errorMessage),
+  TResult when<TResult extends Object?>({
+    required TResult Function() loading,
+    required TResult Function(AccountDetails accountDetails) showAccountDetails,
+    required TResult Function(String errorMessage) error,
   }) {
-    assert(loading != null);
-    assert(showAccountDetails != null);
-    assert(error != null);
     return error(errorMessage);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result loading(),
-    Result showAccountDetails(AccountDetails accountDetails),
-    Result error(String errorMessage),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? loading,
+    TResult Function(AccountDetails accountDetails)? showAccountDetails,
+    TResult Function(String errorMessage)? error,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (error != null) {
       return error(errorMessage);
     }
@@ -390,26 +376,22 @@ class _$Error implements Error {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result loading(Loading value),
-    @required Result showAccountDetails(ShowAccountDetails value),
-    @required Result error(Error value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Loading value) loading,
+    required TResult Function(ShowAccountDetails value) showAccountDetails,
+    required TResult Function(Error value) error,
   }) {
-    assert(loading != null);
-    assert(showAccountDetails != null);
-    assert(error != null);
     return error(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result loading(Loading value),
-    Result showAccountDetails(ShowAccountDetails value),
-    Result error(Error value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Loading value)? loading,
+    TResult Function(ShowAccountDetails value)? showAccountDetails,
+    TResult Function(Error value)? error,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (error != null) {
       return error(this);
     }
@@ -420,6 +402,7 @@ class _$Error implements Error {
 abstract class Error implements MyAddressState {
   const factory Error(String errorMessage) = _$Error;
 
-  String get errorMessage;
-  $ErrorCopyWith<Error> get copyWith;
+  String get errorMessage => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $ErrorCopyWith<Error> get copyWith => throw _privateConstructorUsedError;
 }
