@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'payment_state.dart';
 
@@ -9,28 +9,27 @@ part of 'payment_state.dart';
 
 T _$identity<T>(T value) => value;
 
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 /// @nodoc
 class _$PaymentStateTearOff {
   const _$PaymentStateTearOff();
 
-// ignore: unused_element
   Idle idle() {
     return const Idle();
   }
 
-// ignore: unused_element
   PaymentButtonLoading paymentButtonLoading() {
     return const PaymentButtonLoading();
   }
 
-// ignore: unused_element
   PaymentError paymentError(String errorMessage) {
     return PaymentError(
       errorMessage,
     );
   }
 
-// ignore: unused_element
   PaymentSuccessful paymentSuccessful(PaymentSuccessResponse response) {
     return PaymentSuccessful(
       response,
@@ -39,41 +38,45 @@ class _$PaymentStateTearOff {
 }
 
 /// @nodoc
-// ignore: unused_element
 const $PaymentState = _$PaymentStateTearOff();
 
 /// @nodoc
 mixin _$PaymentState {
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result paymentButtonLoading(),
-    @required Result paymentError(String errorMessage),
-    @required Result paymentSuccessful(PaymentSuccessResponse response),
-  });
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() paymentButtonLoading,
+    required TResult Function(String errorMessage) paymentError,
+    required TResult Function(PaymentSuccessResponse response)
+        paymentSuccessful,
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result paymentButtonLoading(),
-    Result paymentError(String errorMessage),
-    Result paymentSuccessful(PaymentSuccessResponse response),
-    @required Result orElse(),
-  });
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? paymentButtonLoading,
+    TResult Function(String errorMessage)? paymentError,
+    TResult Function(PaymentSuccessResponse response)? paymentSuccessful,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result paymentButtonLoading(PaymentButtonLoading value),
-    @required Result paymentError(PaymentError value),
-    @required Result paymentSuccessful(PaymentSuccessful value),
-  });
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(PaymentButtonLoading value) paymentButtonLoading,
+    required TResult Function(PaymentError value) paymentError,
+    required TResult Function(PaymentSuccessful value) paymentSuccessful,
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result paymentButtonLoading(PaymentButtonLoading value),
-    Result paymentError(PaymentError value),
-    Result paymentSuccessful(PaymentSuccessful value),
-    @required Result orElse(),
-  });
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(PaymentButtonLoading value)? paymentButtonLoading,
+    TResult Function(PaymentError value)? paymentError,
+    TResult Function(PaymentSuccessful value)? paymentSuccessful,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -127,29 +130,25 @@ class _$Idle implements Idle {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result paymentButtonLoading(),
-    @required Result paymentError(String errorMessage),
-    @required Result paymentSuccessful(PaymentSuccessResponse response),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() paymentButtonLoading,
+    required TResult Function(String errorMessage) paymentError,
+    required TResult Function(PaymentSuccessResponse response)
+        paymentSuccessful,
   }) {
-    assert(idle != null);
-    assert(paymentButtonLoading != null);
-    assert(paymentError != null);
-    assert(paymentSuccessful != null);
     return idle();
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result paymentButtonLoading(),
-    Result paymentError(String errorMessage),
-    Result paymentSuccessful(PaymentSuccessResponse response),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? paymentButtonLoading,
+    TResult Function(String errorMessage)? paymentError,
+    TResult Function(PaymentSuccessResponse response)? paymentSuccessful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (idle != null) {
       return idle();
     }
@@ -158,29 +157,24 @@ class _$Idle implements Idle {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result paymentButtonLoading(PaymentButtonLoading value),
-    @required Result paymentError(PaymentError value),
-    @required Result paymentSuccessful(PaymentSuccessful value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(PaymentButtonLoading value) paymentButtonLoading,
+    required TResult Function(PaymentError value) paymentError,
+    required TResult Function(PaymentSuccessful value) paymentSuccessful,
   }) {
-    assert(idle != null);
-    assert(paymentButtonLoading != null);
-    assert(paymentError != null);
-    assert(paymentSuccessful != null);
     return idle(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result paymentButtonLoading(PaymentButtonLoading value),
-    Result paymentError(PaymentError value),
-    Result paymentSuccessful(PaymentSuccessful value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(PaymentButtonLoading value)? paymentButtonLoading,
+    TResult Function(PaymentError value)? paymentError,
+    TResult Function(PaymentSuccessful value)? paymentSuccessful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (idle != null) {
       return idle(this);
     }
@@ -230,29 +224,25 @@ class _$PaymentButtonLoading implements PaymentButtonLoading {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result paymentButtonLoading(),
-    @required Result paymentError(String errorMessage),
-    @required Result paymentSuccessful(PaymentSuccessResponse response),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() paymentButtonLoading,
+    required TResult Function(String errorMessage) paymentError,
+    required TResult Function(PaymentSuccessResponse response)
+        paymentSuccessful,
   }) {
-    assert(idle != null);
-    assert(paymentButtonLoading != null);
-    assert(paymentError != null);
-    assert(paymentSuccessful != null);
     return paymentButtonLoading();
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result paymentButtonLoading(),
-    Result paymentError(String errorMessage),
-    Result paymentSuccessful(PaymentSuccessResponse response),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? paymentButtonLoading,
+    TResult Function(String errorMessage)? paymentError,
+    TResult Function(PaymentSuccessResponse response)? paymentSuccessful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (paymentButtonLoading != null) {
       return paymentButtonLoading();
     }
@@ -261,29 +251,24 @@ class _$PaymentButtonLoading implements PaymentButtonLoading {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result paymentButtonLoading(PaymentButtonLoading value),
-    @required Result paymentError(PaymentError value),
-    @required Result paymentSuccessful(PaymentSuccessful value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(PaymentButtonLoading value) paymentButtonLoading,
+    required TResult Function(PaymentError value) paymentError,
+    required TResult Function(PaymentSuccessful value) paymentSuccessful,
   }) {
-    assert(idle != null);
-    assert(paymentButtonLoading != null);
-    assert(paymentError != null);
-    assert(paymentSuccessful != null);
     return paymentButtonLoading(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result paymentButtonLoading(PaymentButtonLoading value),
-    Result paymentError(PaymentError value),
-    Result paymentSuccessful(PaymentSuccessful value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(PaymentButtonLoading value)? paymentButtonLoading,
+    TResult Function(PaymentError value)? paymentError,
+    TResult Function(PaymentSuccessful value)? paymentSuccessful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (paymentButtonLoading != null) {
       return paymentButtonLoading(this);
     }
@@ -315,7 +300,7 @@ class _$PaymentErrorCopyWithImpl<$Res> extends _$PaymentStateCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object errorMessage = freezed,
+    Object? errorMessage = freezed,
   }) {
     return _then(PaymentError(
       errorMessage == freezed ? _value.errorMessage : errorMessage as String,
@@ -325,7 +310,7 @@ class _$PaymentErrorCopyWithImpl<$Res> extends _$PaymentStateCopyWithImpl<$Res>
 
 /// @nodoc
 class _$PaymentError implements PaymentError {
-  const _$PaymentError(this.errorMessage) : assert(errorMessage != null);
+  const _$PaymentError(this.errorMessage);
 
   @override
   final String errorMessage;
@@ -348,35 +333,32 @@ class _$PaymentError implements PaymentError {
   int get hashCode =>
       runtimeType.hashCode ^ const DeepCollectionEquality().hash(errorMessage);
 
+  @JsonKey(ignore: true)
   @override
   $PaymentErrorCopyWith<PaymentError> get copyWith =>
       _$PaymentErrorCopyWithImpl<PaymentError>(this, _$identity);
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result paymentButtonLoading(),
-    @required Result paymentError(String errorMessage),
-    @required Result paymentSuccessful(PaymentSuccessResponse response),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() paymentButtonLoading,
+    required TResult Function(String errorMessage) paymentError,
+    required TResult Function(PaymentSuccessResponse response)
+        paymentSuccessful,
   }) {
-    assert(idle != null);
-    assert(paymentButtonLoading != null);
-    assert(paymentError != null);
-    assert(paymentSuccessful != null);
     return paymentError(errorMessage);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result paymentButtonLoading(),
-    Result paymentError(String errorMessage),
-    Result paymentSuccessful(PaymentSuccessResponse response),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? paymentButtonLoading,
+    TResult Function(String errorMessage)? paymentError,
+    TResult Function(PaymentSuccessResponse response)? paymentSuccessful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (paymentError != null) {
       return paymentError(errorMessage);
     }
@@ -385,29 +367,24 @@ class _$PaymentError implements PaymentError {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result paymentButtonLoading(PaymentButtonLoading value),
-    @required Result paymentError(PaymentError value),
-    @required Result paymentSuccessful(PaymentSuccessful value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(PaymentButtonLoading value) paymentButtonLoading,
+    required TResult Function(PaymentError value) paymentError,
+    required TResult Function(PaymentSuccessful value) paymentSuccessful,
   }) {
-    assert(idle != null);
-    assert(paymentButtonLoading != null);
-    assert(paymentError != null);
-    assert(paymentSuccessful != null);
     return paymentError(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result paymentButtonLoading(PaymentButtonLoading value),
-    Result paymentError(PaymentError value),
-    Result paymentSuccessful(PaymentSuccessful value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(PaymentButtonLoading value)? paymentButtonLoading,
+    TResult Function(PaymentError value)? paymentError,
+    TResult Function(PaymentSuccessful value)? paymentSuccessful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (paymentError != null) {
       return paymentError(this);
     }
@@ -418,8 +395,10 @@ class _$PaymentError implements PaymentError {
 abstract class PaymentError implements PaymentState {
   const factory PaymentError(String errorMessage) = _$PaymentError;
 
-  String get errorMessage;
-  $PaymentErrorCopyWith<PaymentError> get copyWith;
+  String get errorMessage => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $PaymentErrorCopyWith<PaymentError> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -443,7 +422,7 @@ class _$PaymentSuccessfulCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object response = freezed,
+    Object? response = freezed,
   }) {
     return _then(PaymentSuccessful(
       response == freezed
@@ -455,7 +434,7 @@ class _$PaymentSuccessfulCopyWithImpl<$Res>
 
 /// @nodoc
 class _$PaymentSuccessful implements PaymentSuccessful {
-  const _$PaymentSuccessful(this.response) : assert(response != null);
+  const _$PaymentSuccessful(this.response);
 
   @override
   final PaymentSuccessResponse response;
@@ -478,35 +457,32 @@ class _$PaymentSuccessful implements PaymentSuccessful {
   int get hashCode =>
       runtimeType.hashCode ^ const DeepCollectionEquality().hash(response);
 
+  @JsonKey(ignore: true)
   @override
   $PaymentSuccessfulCopyWith<PaymentSuccessful> get copyWith =>
       _$PaymentSuccessfulCopyWithImpl<PaymentSuccessful>(this, _$identity);
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result paymentButtonLoading(),
-    @required Result paymentError(String errorMessage),
-    @required Result paymentSuccessful(PaymentSuccessResponse response),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() paymentButtonLoading,
+    required TResult Function(String errorMessage) paymentError,
+    required TResult Function(PaymentSuccessResponse response)
+        paymentSuccessful,
   }) {
-    assert(idle != null);
-    assert(paymentButtonLoading != null);
-    assert(paymentError != null);
-    assert(paymentSuccessful != null);
     return paymentSuccessful(response);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result paymentButtonLoading(),
-    Result paymentError(String errorMessage),
-    Result paymentSuccessful(PaymentSuccessResponse response),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? paymentButtonLoading,
+    TResult Function(String errorMessage)? paymentError,
+    TResult Function(PaymentSuccessResponse response)? paymentSuccessful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (paymentSuccessful != null) {
       return paymentSuccessful(response);
     }
@@ -515,29 +491,24 @@ class _$PaymentSuccessful implements PaymentSuccessful {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result paymentButtonLoading(PaymentButtonLoading value),
-    @required Result paymentError(PaymentError value),
-    @required Result paymentSuccessful(PaymentSuccessful value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(PaymentButtonLoading value) paymentButtonLoading,
+    required TResult Function(PaymentError value) paymentError,
+    required TResult Function(PaymentSuccessful value) paymentSuccessful,
   }) {
-    assert(idle != null);
-    assert(paymentButtonLoading != null);
-    assert(paymentError != null);
-    assert(paymentSuccessful != null);
     return paymentSuccessful(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result paymentButtonLoading(PaymentButtonLoading value),
-    Result paymentError(PaymentError value),
-    Result paymentSuccessful(PaymentSuccessful value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(PaymentButtonLoading value)? paymentButtonLoading,
+    TResult Function(PaymentError value)? paymentError,
+    TResult Function(PaymentSuccessful value)? paymentSuccessful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (paymentSuccessful != null) {
       return paymentSuccessful(this);
     }
@@ -549,6 +520,8 @@ abstract class PaymentSuccessful implements PaymentState {
   const factory PaymentSuccessful(PaymentSuccessResponse response) =
       _$PaymentSuccessful;
 
-  PaymentSuccessResponse get response;
-  $PaymentSuccessfulCopyWith<PaymentSuccessful> get copyWith;
+  PaymentSuccessResponse get response => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $PaymentSuccessfulCopyWith<PaymentSuccessful> get copyWith =>
+      throw _privateConstructorUsedError;
 }
