@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'place_order_state.dart';
 
@@ -9,26 +9,25 @@ part of 'place_order_state.dart';
 
 T _$identity<T>(T value) => value;
 
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 /// @nodoc
 class _$PlaceOrderStateTearOff {
   const _$PlaceOrderStateTearOff();
 
-// ignore: unused_element
   Idle idle() {
     return const Idle();
   }
 
-// ignore: unused_element
   OrderPlacedInProgress orderPlacedInProgress() {
     return const OrderPlacedInProgress();
   }
 
-// ignore: unused_element
   OrderSuccessfullyPlaced orderSuccessfullyPlaced() {
     return const OrderSuccessfullyPlaced();
   }
 
-// ignore: unused_element
   OrderNotPlaced orderNotPlaced(String message) {
     return OrderNotPlaced(
       message,
@@ -37,41 +36,46 @@ class _$PlaceOrderStateTearOff {
 }
 
 /// @nodoc
-// ignore: unused_element
 const $PlaceOrderState = _$PlaceOrderStateTearOff();
 
 /// @nodoc
 mixin _$PlaceOrderState {
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result orderPlacedInProgress(),
-    @required Result orderSuccessfullyPlaced(),
-    @required Result orderNotPlaced(String message),
-  });
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() orderPlacedInProgress,
+    required TResult Function() orderSuccessfullyPlaced,
+    required TResult Function(String message) orderNotPlaced,
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result orderPlacedInProgress(),
-    Result orderSuccessfullyPlaced(),
-    Result orderNotPlaced(String message),
-    @required Result orElse(),
-  });
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? orderPlacedInProgress,
+    TResult Function()? orderSuccessfullyPlaced,
+    TResult Function(String message)? orderNotPlaced,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result orderPlacedInProgress(OrderPlacedInProgress value),
-    @required Result orderSuccessfullyPlaced(OrderSuccessfullyPlaced value),
-    @required Result orderNotPlaced(OrderNotPlaced value),
-  });
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(OrderPlacedInProgress value)
+        orderPlacedInProgress,
+    required TResult Function(OrderSuccessfullyPlaced value)
+        orderSuccessfullyPlaced,
+    required TResult Function(OrderNotPlaced value) orderNotPlaced,
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result orderPlacedInProgress(OrderPlacedInProgress value),
-    Result orderSuccessfullyPlaced(OrderSuccessfullyPlaced value),
-    Result orderNotPlaced(OrderNotPlaced value),
-    @required Result orElse(),
-  });
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(OrderPlacedInProgress value)? orderPlacedInProgress,
+    TResult Function(OrderSuccessfullyPlaced value)? orderSuccessfullyPlaced,
+    TResult Function(OrderNotPlaced value)? orderNotPlaced,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -126,29 +130,24 @@ class _$Idle implements Idle {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result orderPlacedInProgress(),
-    @required Result orderSuccessfullyPlaced(),
-    @required Result orderNotPlaced(String message),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() orderPlacedInProgress,
+    required TResult Function() orderSuccessfullyPlaced,
+    required TResult Function(String message) orderNotPlaced,
   }) {
-    assert(idle != null);
-    assert(orderPlacedInProgress != null);
-    assert(orderSuccessfullyPlaced != null);
-    assert(orderNotPlaced != null);
     return idle();
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result orderPlacedInProgress(),
-    Result orderSuccessfullyPlaced(),
-    Result orderNotPlaced(String message),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? orderPlacedInProgress,
+    TResult Function()? orderSuccessfullyPlaced,
+    TResult Function(String message)? orderNotPlaced,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (idle != null) {
       return idle();
     }
@@ -157,29 +156,26 @@ class _$Idle implements Idle {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result orderPlacedInProgress(OrderPlacedInProgress value),
-    @required Result orderSuccessfullyPlaced(OrderSuccessfullyPlaced value),
-    @required Result orderNotPlaced(OrderNotPlaced value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(OrderPlacedInProgress value)
+        orderPlacedInProgress,
+    required TResult Function(OrderSuccessfullyPlaced value)
+        orderSuccessfullyPlaced,
+    required TResult Function(OrderNotPlaced value) orderNotPlaced,
   }) {
-    assert(idle != null);
-    assert(orderPlacedInProgress != null);
-    assert(orderSuccessfullyPlaced != null);
-    assert(orderNotPlaced != null);
     return idle(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result orderPlacedInProgress(OrderPlacedInProgress value),
-    Result orderSuccessfullyPlaced(OrderSuccessfullyPlaced value),
-    Result orderNotPlaced(OrderNotPlaced value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(OrderPlacedInProgress value)? orderPlacedInProgress,
+    TResult Function(OrderSuccessfullyPlaced value)? orderSuccessfullyPlaced,
+    TResult Function(OrderNotPlaced value)? orderNotPlaced,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (idle != null) {
       return idle(this);
     }
@@ -229,29 +225,24 @@ class _$OrderPlacedInProgress implements OrderPlacedInProgress {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result orderPlacedInProgress(),
-    @required Result orderSuccessfullyPlaced(),
-    @required Result orderNotPlaced(String message),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() orderPlacedInProgress,
+    required TResult Function() orderSuccessfullyPlaced,
+    required TResult Function(String message) orderNotPlaced,
   }) {
-    assert(idle != null);
-    assert(orderPlacedInProgress != null);
-    assert(orderSuccessfullyPlaced != null);
-    assert(orderNotPlaced != null);
     return orderPlacedInProgress();
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result orderPlacedInProgress(),
-    Result orderSuccessfullyPlaced(),
-    Result orderNotPlaced(String message),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? orderPlacedInProgress,
+    TResult Function()? orderSuccessfullyPlaced,
+    TResult Function(String message)? orderNotPlaced,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (orderPlacedInProgress != null) {
       return orderPlacedInProgress();
     }
@@ -260,29 +251,26 @@ class _$OrderPlacedInProgress implements OrderPlacedInProgress {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result orderPlacedInProgress(OrderPlacedInProgress value),
-    @required Result orderSuccessfullyPlaced(OrderSuccessfullyPlaced value),
-    @required Result orderNotPlaced(OrderNotPlaced value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(OrderPlacedInProgress value)
+        orderPlacedInProgress,
+    required TResult Function(OrderSuccessfullyPlaced value)
+        orderSuccessfullyPlaced,
+    required TResult Function(OrderNotPlaced value) orderNotPlaced,
   }) {
-    assert(idle != null);
-    assert(orderPlacedInProgress != null);
-    assert(orderSuccessfullyPlaced != null);
-    assert(orderNotPlaced != null);
     return orderPlacedInProgress(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result orderPlacedInProgress(OrderPlacedInProgress value),
-    Result orderSuccessfullyPlaced(OrderSuccessfullyPlaced value),
-    Result orderNotPlaced(OrderNotPlaced value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(OrderPlacedInProgress value)? orderPlacedInProgress,
+    TResult Function(OrderSuccessfullyPlaced value)? orderSuccessfullyPlaced,
+    TResult Function(OrderNotPlaced value)? orderNotPlaced,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (orderPlacedInProgress != null) {
       return orderPlacedInProgress(this);
     }
@@ -332,29 +320,24 @@ class _$OrderSuccessfullyPlaced implements OrderSuccessfullyPlaced {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result orderPlacedInProgress(),
-    @required Result orderSuccessfullyPlaced(),
-    @required Result orderNotPlaced(String message),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() orderPlacedInProgress,
+    required TResult Function() orderSuccessfullyPlaced,
+    required TResult Function(String message) orderNotPlaced,
   }) {
-    assert(idle != null);
-    assert(orderPlacedInProgress != null);
-    assert(orderSuccessfullyPlaced != null);
-    assert(orderNotPlaced != null);
     return orderSuccessfullyPlaced();
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result orderPlacedInProgress(),
-    Result orderSuccessfullyPlaced(),
-    Result orderNotPlaced(String message),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? orderPlacedInProgress,
+    TResult Function()? orderSuccessfullyPlaced,
+    TResult Function(String message)? orderNotPlaced,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (orderSuccessfullyPlaced != null) {
       return orderSuccessfullyPlaced();
     }
@@ -363,29 +346,26 @@ class _$OrderSuccessfullyPlaced implements OrderSuccessfullyPlaced {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result orderPlacedInProgress(OrderPlacedInProgress value),
-    @required Result orderSuccessfullyPlaced(OrderSuccessfullyPlaced value),
-    @required Result orderNotPlaced(OrderNotPlaced value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(OrderPlacedInProgress value)
+        orderPlacedInProgress,
+    required TResult Function(OrderSuccessfullyPlaced value)
+        orderSuccessfullyPlaced,
+    required TResult Function(OrderNotPlaced value) orderNotPlaced,
   }) {
-    assert(idle != null);
-    assert(orderPlacedInProgress != null);
-    assert(orderSuccessfullyPlaced != null);
-    assert(orderNotPlaced != null);
     return orderSuccessfullyPlaced(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result orderPlacedInProgress(OrderPlacedInProgress value),
-    Result orderSuccessfullyPlaced(OrderSuccessfullyPlaced value),
-    Result orderNotPlaced(OrderNotPlaced value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(OrderPlacedInProgress value)? orderPlacedInProgress,
+    TResult Function(OrderSuccessfullyPlaced value)? orderSuccessfullyPlaced,
+    TResult Function(OrderNotPlaced value)? orderNotPlaced,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (orderSuccessfullyPlaced != null) {
       return orderSuccessfullyPlaced(this);
     }
@@ -418,7 +398,7 @@ class _$OrderNotPlacedCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object message = freezed,
+    Object? message = freezed,
   }) {
     return _then(OrderNotPlaced(
       message == freezed ? _value.message : message as String,
@@ -428,7 +408,7 @@ class _$OrderNotPlacedCopyWithImpl<$Res>
 
 /// @nodoc
 class _$OrderNotPlaced implements OrderNotPlaced {
-  const _$OrderNotPlaced(this.message) : assert(message != null);
+  const _$OrderNotPlaced(this.message);
 
   @override
   final String message;
@@ -450,35 +430,31 @@ class _$OrderNotPlaced implements OrderNotPlaced {
   int get hashCode =>
       runtimeType.hashCode ^ const DeepCollectionEquality().hash(message);
 
+  @JsonKey(ignore: true)
   @override
   $OrderNotPlacedCopyWith<OrderNotPlaced> get copyWith =>
       _$OrderNotPlacedCopyWithImpl<OrderNotPlaced>(this, _$identity);
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result orderPlacedInProgress(),
-    @required Result orderSuccessfullyPlaced(),
-    @required Result orderNotPlaced(String message),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() orderPlacedInProgress,
+    required TResult Function() orderSuccessfullyPlaced,
+    required TResult Function(String message) orderNotPlaced,
   }) {
-    assert(idle != null);
-    assert(orderPlacedInProgress != null);
-    assert(orderSuccessfullyPlaced != null);
-    assert(orderNotPlaced != null);
     return orderNotPlaced(message);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result orderPlacedInProgress(),
-    Result orderSuccessfullyPlaced(),
-    Result orderNotPlaced(String message),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? orderPlacedInProgress,
+    TResult Function()? orderSuccessfullyPlaced,
+    TResult Function(String message)? orderNotPlaced,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (orderNotPlaced != null) {
       return orderNotPlaced(message);
     }
@@ -487,29 +463,26 @@ class _$OrderNotPlaced implements OrderNotPlaced {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result orderPlacedInProgress(OrderPlacedInProgress value),
-    @required Result orderSuccessfullyPlaced(OrderSuccessfullyPlaced value),
-    @required Result orderNotPlaced(OrderNotPlaced value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(OrderPlacedInProgress value)
+        orderPlacedInProgress,
+    required TResult Function(OrderSuccessfullyPlaced value)
+        orderSuccessfullyPlaced,
+    required TResult Function(OrderNotPlaced value) orderNotPlaced,
   }) {
-    assert(idle != null);
-    assert(orderPlacedInProgress != null);
-    assert(orderSuccessfullyPlaced != null);
-    assert(orderNotPlaced != null);
     return orderNotPlaced(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result orderPlacedInProgress(OrderPlacedInProgress value),
-    Result orderSuccessfullyPlaced(OrderSuccessfullyPlaced value),
-    Result orderNotPlaced(OrderNotPlaced value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(OrderPlacedInProgress value)? orderPlacedInProgress,
+    TResult Function(OrderSuccessfullyPlaced value)? orderSuccessfullyPlaced,
+    TResult Function(OrderNotPlaced value)? orderNotPlaced,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (orderNotPlaced != null) {
       return orderNotPlaced(this);
     }
@@ -520,6 +493,8 @@ class _$OrderNotPlaced implements OrderNotPlaced {
 abstract class OrderNotPlaced implements PlaceOrderState {
   const factory OrderNotPlaced(String message) = _$OrderNotPlaced;
 
-  String get message;
-  $OrderNotPlacedCopyWith<OrderNotPlaced> get copyWith;
+  String get message => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $OrderNotPlacedCopyWith<OrderNotPlaced> get copyWith =>
+      throw _privateConstructorUsedError;
 }
