@@ -4,7 +4,7 @@ class WidgetVisibility extends StatelessWidget {
   final bool visible;
   final Widget child;
 
-  WidgetVisibility({@required this.visible, @required this.child})
+  WidgetVisibility({required this.visible, required this.child})
       : assert(visible != null && child != null);
 
   @override
