@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:fluttercommerce/src/res/app_theme.dart';
-import 'package:fluttercommerce/src/routes/router.gr.dart';
+import 'package:fluttercommerce/src/routes/routers.dart';
 import 'package:fluttercommerce/src/ui/screens/splash_screen.dart';
 
 class App extends StatefulWidget {
@@ -14,7 +14,7 @@ class _AppState extends State<App> {
     return MaterialApp(
       theme: AppTheme.appTheme(),
       debugShowCheckedModeBanner: false,
-      onGenerateRoute: AppRouter(),
+      onGenerateRoute: AppRouter() as Route<dynamic>? Function(RouteSettings)?,
       initialRoute: Routes.splashScreen,
       home: SplashScreen(),
     );
