@@ -10,12 +10,12 @@ class CartStatusProvider with ChangeNotifier {
     num price = 0;
     List.generate(cartItems.length, (index) {
       price =
-          price + (cartItems[index].currentPrice * cartItems[index].numOfItems);
+          price + (cartItems[index].currentPrice! * cartItems[index].numOfItems!);
     });
     return price;
   }
 
-  String get currency => noOfItemsInCart > 0 ? cartItems[0]?.currency : "";
+  String? get currency => noOfItemsInCart > 0 ? cartItems[0]?.currency : "";
 
   List<CartModel> get cartItems => _cartItems;
 
