@@ -6,7 +6,7 @@ import 'package:fluttercommerce/src/di/app_injector.dart';
 import 'package:fluttercommerce/src/res/string_constants.dart';
 import 'package:fluttercommerce/src/res/styles.dart';
 import 'package:fluttercommerce/src/res/text_styles.dart';
-import 'package:fluttercommerce/src/routes/router.gr.dart';
+import 'package:fluttercommerce/src/routes/routers.dart';
 import 'package:fluttercommerce/src/ui/common/commom_text_field.dart';
 import 'package:fluttercommerce/src/ui/common/common_app_loader.dart';
 import 'package:fluttercommerce/src/ui/common/common_button.dart';
@@ -14,7 +14,7 @@ import 'package:fluttercommerce/src/ui/common/common_button.dart';
 class OTPLoginScreen extends StatefulWidget {
   final String phoneNumber;
 
-  OTPLoginScreen({this.phoneNumber});
+  OTPLoginScreen(this.phoneNumber);
 
   @override
   _OTPLoginScreenState createState() => _OTPLoginScreenState();
@@ -29,10 +29,10 @@ class _OTPLoginScreenState extends State<OTPLoginScreen> {
   @override
   void initState() {
     super.initState();
-    otpLoginCubit.sendOtp(widget.phoneNumber);
+    otpLoginCubit!.sendOtp(widget.phoneNumber);
     otpNumberController.addListener(() {
       print(otpNumberController.text);
-      otpLoginCubit.validateButton(otpNumberController.text);
+      otpLoginCubit!.validateButton(otpNumberController.text);
     });
   }
 
@@ -121,7 +121,7 @@ class _OTPLoginScreenState extends State<OTPLoginScreen> {
                 height: 20,
               ),
               BlocConsumer<OtpLoginCubit, OtpLoginState>(
-                cubit: otpLoginCubit,
+                cubit: otpLoginCubit!,
                 listener: (BuildContext context, OtpLoginState state) {
                   state.when(
                       idle: () {},
@@ -171,7 +171,7 @@ class _OTPLoginScreenState extends State<OTPLoginScreen> {
                 height: 20,
               ),
               BlocConsumer<OtpLoginCubit, OtpLoginState>(
-                cubit: otpLoginCubit,
+                cubit: otpLoginCubit!,
                 listener: (BuildContext context, OtpLoginState state) {},
                 builder: (BuildContext context, OtpLoginState state) {
                   if (state is ResendOtpLoading) return CommonAppLoader();
@@ -204,8 +204,8 @@ class _OTPLoginScreenState extends State<OTPLoginScreen> {
   }
 
   void onButtonTap() {
-    if (_formKey.currentState.validate()) {
-      otpLoginCubit.loginWithOtp(otpNumberController.text.trim());
+    if (_formKey.currentState!.validate()) {
+      otpLoginCubit!.loginWithOtp(otpNumberController.text.trim());
     }
   }
 }
